Refactoring app. Now every entity has it's own directory with views, models and urls file
